+++ Plugin/src/Patches/PlayerControllerBPatch.cs
@@ -1,4 +1,6 @@
 using CodeRebirth.MapStuff;
+using CodeRebirth.Util.PlayerManager;
+using GameNetcodeStuff;
 using HarmonyLib;
 using Unity.Netcode;
 using UnityEngine;
@@ -7,10 +9,9 @@ namespace CodeRebirth.Patches;
 
 [HarmonyPatch(typeof(GameNetcodeStuff.PlayerControllerB))]
 static class PlayerControllerBPatch {
-    public static bool mountedPlayer = false;
 	[HarmonyPatch(nameof(GameNetcodeStuff.PlayerControllerB.PlayFootstepSound)), HarmonyPrefix]
-	public static bool PlayFootstepSound() {
-        if (mountedPlayer) {
+	public static bool PlayFootstepSound(PlayerControllerB __instance) {
+        if (__instance.gameObject.GetComponent<CodeRebirthPlayerManager>().ridingHoverboard) {
             return false;
         } else {
             return true;