+++ Plugin/src/WeatherStuff/CodeRebirthWeathers.cs
@@ -68,9 +68,9 @@ public class CodeRebirthWeathers : MonoBehaviour {
 			}
 			position += new Vector3(random.NextFloat(minX, maxX), random.NextFloat(minY, maxY), random.NextFloat(minZ, maxZ));
 			position = RoundManager.Instance.GetRandomNavMeshPositionInBoxPredictable(pos: position, radius: radius, randomSeed: random);
-			if (!Plugin.ModConfig.ConfigMeteorHitShip.Value && Vector3.Distance(position, StartOfRound.Instance.shipBounds.transform.position) < 8) {
+			if (!Plugin.ModConfig.ConfigMeteorHitShip.Value && Vector3.Distance(position, StartOfRound.Instance.shipBounds.transform.position) < 13) {
 				for (int i = 0; i < 5; i++) {
-					if (Vector3.Distance(position, StartOfRound.Instance.shipBounds.transform.position) < 8) {
+					if (Vector3.Distance(position, StartOfRound.Instance.shipBounds.transform.position) < 13) {
 						nextNode = random.NextItem(nodes);
 						position = nextNode.transform.position;
 						position += new Vector3(random.NextFloat(minX, maxX), random.NextFloat(minY, maxY), random.NextFloat(minZ, maxZ));